pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".

[tool.poetry]
name = "aws-cdk-python-starterkit"
version = "0.1.0"
description = "Create and deploy an AWS CDK app on your AWS account in less than 5 minutes using GitHub actions!"
authors = [ "Danny Steenman <danny@towardsthecloud.com>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  aws-cdk-github-oidc = "*"
  aws-cdk-lib = "2.133.0"
  constructs = "^10.0.5"
  python = "^3.8"

[tool.poetry.group.dev.dependencies]
projen = "0.80.12"
ruff = "*"
pytest = "7.4.3"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
